Fix errors on deprecated NumPy `np.int` and `np.float`
NumPy aliases for generic data types have been deprecated and now give an error as of NumPy v1.24 (see https://numpy.org/doc/stable/release/1.24.0-notes.html#expired-deprecations). Change these references to the standard Python types (see https://numpy.org/doc/stable/release/1.20.0-notes.html#deprecations).

+++ magmap/gui/visualizer.py
@@ -455,10 +455,10 @@ class Visualization(HasTraits):
     _import_btn = Button("Import files")
     _import_btn_enabled = Bool
     _import_clear_btn = Button("Clear import")
-    _import_res = Array(np.float, shape=(1, 3))
+    _import_res = Array(float, shape=(1, 3))
     _import_mag = Float(1.0)
     _import_zoom = Float(1.0)
-    _import_shape = Array(np.int, shape=(1, 5), editor=ArrayEditor(
+    _import_shape = Array(int, shape=(1, 5), editor=ArrayEditor(
         width=-40, format_str="%0d"))
     # map bits to bytes for constructing Numpy data type
     _IMPORT_BITS = OrderedDict((
@@ -1691,7 +1691,7 @@ class Visualization(HasTraits):
             # get atlas label at ROI center
             center = np.add(
                 curr_offset, 
-                np.around(np.divide(curr_roi_size, 2)).astype(np.int))
+                np.around(np.divide(curr_roi_size, 2)).astype(int))
             self._atlas_label = ontology.get_label(
                 center[::-1], config.labels_img, config.labels_ref.ref_lookup, 
                 config.labels_scaling, level, rounding=True)
@@ -2641,7 +2641,7 @@ class Visualization(HasTraits):
                     # get colors of corresponding atlas labels
                     if coords is None: return None
                     blob_ids = ontology.get_label_ids_from_position(
-                        coords.astype(np.int), config.labels_img)
+                        coords.astype(int), config.labels_img)
                     atlas_cmap = config.cmap_labels(
                         config.cmap_labels.convert_img_labels(blob_ids))
                     atlas_cmap[:, :3] *= 255
@@ -2660,7 +2660,7 @@ class Visualization(HasTraits):
             else:
                 # default to color by channel
                 blob_chls = detector.Blobs.get_blobs_channel(
-                    segs_all[self.segs_in_mask]).astype(np.int)
+                    segs_all[self.segs_in_mask]).astype(int)
                 cmap = colormaps.discrete_colormap(
                     max(blob_chls) + 1, alpha, True, config.seed)
                 self.segs_cmap = cmap[blob_chls]
@@ -3421,7 +3421,7 @@ class Visualization(HasTraits):
             curr_roi_size = self.roi_array[0].astype(int)
             corner = np.subtract(
                 centroid, 
-                np.around(np.divide(curr_roi_size[::-1], 2)).astype(np.int))
+                np.around(np.divide(curr_roi_size[::-1], 2)).astype(int))
             self.z_offset, self.y_offset, self.x_offset = corner
             self._check_roi_position()
             self.show_3d()
@@ -3998,7 +3998,7 @@ class Visualization(HasTraits):
         self._import_paths = []
         self._import_mode = None
         self._import_res = np.ones((1, 3))
-        self._import_shape = np.zeros((1, 5), dtype=np.int)
+        self._import_shape = np.zeros((1, 5), dtype=int)
         self._import_bit = [tuple(self._IMPORT_BITS.keys())[0]]
         self._import_data_type = [tuple(self._IMPORT_DATA_TYPES.keys())[0]]
         self._import_byte_order = [tuple(self._IMPORT_BYTE_ORDERS.keys())[0]]
